consolidated utils until proven otherwise
